Don't Reset next_occurrence For Cloned Objects (#108)
Amend logic determining if object is new by considering last_occurrence so that cloned objects will not be treated as new and have their next_occurrence reset to the first date in the series. Expanded clone_with_day_offset to copy of the source object's next_occurrence, with offset, so that cloned object will not have to processed stale next_occurrence dates to 'catch up'. Added test to flex.